Fixed home screen shortcuts
